Make config parsing more readable, put server logic into own file and put server initialization into own function
