+++ Cargo.toml
@@ -3,6 +3,9 @@ name = "mjpeg-proxy"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "mjpeg_proxy"
+
 [dependencies]
 futures = "0.3.31"
 mime = "0.3.17"
@@ -12,5 +15,6 @@ hyper = { version = "0.14", features = ["http1", "client", "server", "stream", "
 lazy-static-include = "3.2.1"
 tokio-stream = { version = "0.1.17", features = ["sync"] }
 lazy_static = "1.5.0"
+serde = { version = "1.0.217", features = ["derive"] }
 serde_yaml = "0.9.34"
 chrono = "0.4.39"
